# Conflicts:
#	FWLog.Data/FWLog.Data.csproj
#	FWLog.Data/Repository/GeneralCtx/PerfilUsuarioRepository.cs
#	FWLog.Data/UnitOfWork.cs

+++ FWLog.Data/Repository/GeneralCtx/LoteRepository.cs
@@ -19,7 +19,7 @@ namespace FWLog.Data.Repository.GeneralCtx
             return Entities.Lote.FirstOrDefault(f => f.IdNotaFiscal == idNotaFiscal);
         }
 
-        public IEnumerable<Lote> Obter(int idEmpresa)
+        public IEnumerable<Lote> Obter(long idEmpresa)
         {
             IEnumerable<Lote> lote = null;
